+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Types of changes can be: Added/Changed/Deprecated/Removed/Fixed/Security
 
 ## [UNRELEASED]
 ### Added
+- open files inside app (jpg, .png, .gif, .mp3, .mp4, .mov, .wav, .txt, .md)
 - edit profile information and privacy settings
 
 ### Changed